Refresh wallet after background
