+++ navigation/AppContainer.tsx
@@ -2,19 +2,65 @@ import * as React from 'react';
 import {NavigationContainer} from '@react-navigation/native';
 import createStackNavigator from './createStackNavigator';
 import {RootStackParamList} from './type';
-import Text from '../components/Text';
+import {AppState} from 'react-native';
 import Intro from './Intro';
 import CreateNewWallet from './CreateNewWallet';
-import {View} from 'react-native';
 import Loading from '../components/Loading';
 import OnBoarding from '../screens/OnBoarding';
 import Wallet from '../screens/Wallet';
 import OpenWallet from './OpenWallet';
 import ImportWallet from './ImportWallet';
+import useLockedScreen from '../hooks/useLockedScreen';
+import useAsyncStorage from '../hooks/useAsyncStorage';
+import useWallet from '../hooks/useWallet';
+import {useEffect, useRef, useState} from 'react';
+import LocalAuth from '../utils/LocalAuth';
+import {BlurView} from 'expo';
 
 const Stack = createStackNavigator<RootStackParamList>();
 
 const AppContainer = (props: any) => {
+  const {lockedScreen, setLockedScreen} = useLockedScreen();
+  const [lockAfterBackground, setLockAfterBackground] = useAsyncStorage(
+    'lockAfterBackground',
+    'false',
+  );
+
+  const {refreshWallet} = useWallet();
+
+  const appState = useRef(AppState.currentState);
+
+  const [appStateVisible, setAppStateVisible] = useState(appState.current);
+
+  useEffect(() => {
+    if (appStateVisible == 'active' && lockedScreen) {
+      LocalAuth((error: any) => {
+        if (!error) {
+          setLockedScreen(false);
+        } else {
+          setLockedScreen(true);
+        }
+      });
+    }
+  }, [appStateVisible]);
+
+  useEffect(() => {
+    const subscription = AppState.addEventListener('change', nextAppState => {
+      if (appState.current.match(/background/) && nextAppState === 'active') {
+        if (refreshWallet) refreshWallet();
+        if (lockAfterBackground === 'true') {
+          setLockedScreen(true);
+        }
+      }
+      appState.current = nextAppState;
+      setAppStateVisible(appState.current);
+    });
+
+    return () => {
+      subscription.remove();
+    };
+  }, [refreshWallet, lockAfterBackground, setLockedScreen]);
+
   return props.loaded ? (
     <NavigationContainer>
       <Stack.Navigator
